backend/requirements.txt
# CropCure Backend Dependencies
# Core Flask framework
Flask==2.3.3
flask-cors==4.0.0

# Machine Learning
tensorflow==2.19.0
numpy==1.26.3

# Image Processing
Pillow==10.0.1

# AI Integration
openai==0.28.0

# Development and Production
gunicorn==21.2.0
python-dotenv==1.0.0

# Optional: For better performance
# tensorflow-gpu==2.13.0  # Uncomment if you have GPU support
